Partially hardened `delete_db` to only work if `create_db` is called. Also should fail now if the db is already existing.

+++ tortoise/backends/sqlite/client.py
@@ -26,10 +26,11 @@ class SqliteClient(BaseDBAsyncClient):
         )
 
     async def create_connection(self):
-        pass
+        if self.create_db and os.path.isfile(self.filename):
+            raise OperationalError('DB cannot be created, as it already exists.')
 
     async def close(self):
-        if self.delete_db:
+        if self.delete_db and self.create_db:
             try:
                 os.remove(self.filename)
             except FileNotFoundError:
